Remove members mod from the membership
- move all the code from the members mod to the lib.rs
- update all dependencies
